chore: consolidate security audit utilities
